Remove unnecessary object storage code, part 1
