--- astra/src/main/java/com/slack/astra/blobfs/BlobFsConfig.java
@@ -1,5 +1,0 @@
-package com.slack.astra.blobfs;
-
-public interface BlobFsConfig {
-  String getProperty(String propertyName);
-}
