+++ astra/src/main/java/com/slack/astra/blobfs/BlobFsUtils.java
@@ -1,17 +1,16 @@
-package com.slack.astra.logstore;
+package com.slack.astra.blobfs;
 
-import com.slack.astra.blobfs.BlobFs;
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** This class contains static methods that help with blobfs operations. */
+@Deprecated
 public class BlobFsUtils {
 
   private static final Logger LOG = LoggerFactory.getLogger(BlobFsUtils.class);
@@ -20,6 +19,7 @@ public class BlobFsUtils {
   public static final String DELIMITER = "/";
   public static final String FILE_FORMAT = "%s://%s/%s";
 
+  @Deprecated
   public static int copyToS3(
       Path sourceDirPath, Collection<String> files, String bucket, String prefix, BlobFs blobFs)
       throws Exception {
@@ -35,6 +35,7 @@ public class BlobFsUtils {
     return success;
   }
 
+  @Deprecated
   public static URI createURI(String bucket, String prefix, String fileName) {
     return (prefix != null && !prefix.isEmpty())
         ? URI.create(String.format(FILE_FORMAT, SCHEME, bucket + DELIMITER + prefix, fileName))
@@ -43,6 +44,7 @@ public class BlobFsUtils {
 
   // TODO: Can we copy files without list files and a prefix only?
   // TODO: Take a complete URI as this is the format stored in snapshot data
+  @Deprecated
   public static String[] copyFromS3(
       String bucket, String prefix, BlobFs s3BlobFs, Path localDirPath) throws Exception {
     LOG.debug("Copying files from bucket={} prefix={} using directory", bucket, prefix);
@@ -53,15 +55,5 @@ public class BlobFsUtils {
         .map(File::toString)
         .distinct()
         .toArray(String[]::new);
-  }
-
-  public static void copyToLocalPath(
-      Path sourceDirPath, Collection<String> files, Path destDirPath, BlobFs blobFs)
-      throws IOException {
-    for (String file : files) {
-      blobFs.copy(
-          Paths.get(sourceDirPath.toAbsolutePath().toString(), file).toUri(),
-          Paths.get(destDirPath.toAbsolutePath().toString(), file).toUri());
-    }
   }
 }
