+++ astra/src/test/java/com/slack/astra/chunk/ReadOnlyChunkImplTest.java
@@ -1,9 +1,9 @@
 package com.slack.astra.chunk;
 
+import static com.slack.astra.blobfs.BlobFsUtils.copyToS3;
 import static com.slack.astra.chunk.ReadOnlyChunkImpl.CHUNK_ASSIGNMENT_TIMER;
 import static com.slack.astra.chunk.ReadOnlyChunkImpl.CHUNK_EVICTION_TIMER;
 import static com.slack.astra.chunk.ReadWriteChunk.SCHEMA_FILE_NAME;
-import static com.slack.astra.logstore.BlobFsUtils.copyToS3;
 import static com.slack.astra.logstore.LuceneIndexStoreImpl.COMMITS_TIMER;
 import static com.slack.astra.logstore.LuceneIndexStoreImpl.MESSAGES_FAILED_COUNTER;
 import static com.slack.astra.logstore.LuceneIndexStoreImpl.MESSAGES_RECEIVED_COUNTER;
@@ -18,8 +18,7 @@ import static org.awaitility.Awaitility.await;
 import brave.Tracing;
 import com.adobe.testing.s3mock.junit5.S3MockExtension;
 import com.slack.astra.blobfs.ChunkStore;
-import com.slack.astra.blobfs.LocalBlobFs;
-import com.slack.astra.blobfs.s3.S3CrtBlobFs;
+import com.slack.astra.blobfs.S3CrtBlobFs;
 import com.slack.astra.blobfs.s3.S3TestUtils;
 import com.slack.astra.logstore.LogMessage;
 import com.slack.astra.logstore.LuceneIndexStoreImpl;
@@ -678,11 +677,7 @@ public class ReadOnlyChunkImplTest {
     IndexCommit indexCommit = logStore.getIndexCommit();
     filesToUpload.addAll(indexCommit.getFileNames());
 
-    LocalBlobFs localBlobFs = new LocalBlobFs();
-
-    logStore.close();
-    assertThat(localBlobFs.listFiles(dirPath.toUri(), false).length)
-        .isGreaterThanOrEqualTo(filesToUpload.size());
+    assertThat(dirPath.toFile().listFiles().length).isGreaterThanOrEqualTo(filesToUpload.size());
 
     // Copy files to S3.
     copyToS3(dirPath, filesToUpload, TEST_S3_BUCKET, snapshotId, s3CrtBlobFs);
