+++ astra/src/test/java/com/slack/astra/recovery/RecoveryServiceTest.java
@@ -25,9 +25,9 @@ import brave.Tracing;
 import com.adobe.testing.s3mock.junit5.S3MockExtension;
 import com.google.common.collect.Maps;
 import com.slack.astra.blobfs.BlobFs;
-import com.slack.astra.blobfs.s3.S3CrtBlobFs;
+import com.slack.astra.blobfs.BlobFsUtils;
+import com.slack.astra.blobfs.S3CrtBlobFs;
 import com.slack.astra.blobfs.s3.S3TestUtils;
-import com.slack.astra.logstore.BlobFsUtils;
 import com.slack.astra.metadata.core.AstraMetadataTestUtils;
 import com.slack.astra.metadata.core.CuratorBuilder;
 import com.slack.astra.metadata.recovery.RecoveryNodeMetadata;
